Refactor package name of the EmptyBodyInterceptor
